Update index.html
Now more infromation is saved and retrieved from the JSON this operation is backwards compatible. Like language and so on. 

Now we have global margin to increase and decrease all prices. 
Ability to see original prices and currency.
Ability to edit them.
Ability to add margin to specific items and subitems.
Ability to have multiple images on one line.
Ability to move item to subitem and back and then to optional and back.

Ability to copy the json withouth the images to clipboard straight away for copy pasting fast

Ability to make text bold and italic by ctrl + b and ctrl + i and also by selecting text and pressing B or I. 

It is no longer possible to edit the total price though. 
Currently we test the fact that the subitems are idented

TODO: remove items button needs to be added
ability to add ability to track if items are removed like when the optional items part completely gone